python_files/requirements.txt
opencv-python
pillow
google.generativeai
easyocr
spacy
scikit-image
requests
tensorflow 
numpy
firebase-admin
datetime
picamera
scikit-image
argparse
geopy
pyttsx3
mediapipe
face-recognition
langchain
google-auth-oauthlib
langchain-google-genai
soundfile
torch
torchaudio
transformers
pyaudio
webrtcvad
serial
pynmea2
python-gsmmodem-new
gpiod
mpu6050-raspberrypi
SpeechRecognition
--extra-index-url https://download.pytorch.org/whl/cu116
git+https://github.com/openai/whisper.git
https://github.com/kpu/kenlm/archive/master.zip